# backend/openapi.json
{
  "components": {
    "schemas": {
      "AccountBalance": {
        "properties": {
          "balance_minor": {
            "title": "Balance Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "currency",
          "balance_minor"
        ],
        "title": "AccountBalance",
        "type": "object"
      },
...
      "ExpireApprovals": {
        "additionalProperties": false,
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "ExpireApprovals",
        "type": "object"
      },
      "FundingFailure": {
        "additionalProperties": false,
        "properties": {
          "member_id": {
            "title": "Member Id",
            "type": "string"
          },
          "unable_to_fund": {
            "default": true,
            "title": "Unable To Fund",
            "type": "boolean"
          }
        },
        "required": [
          "member_id"
        ],
        "title": "FundingFailure",
        "type": "object"
      },
      "FundingState": {
        "properties": {
          "funding_blocked": {
            "title": "Funding Blocked",
            "type": "boolean"
          },
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "member_id",
          "name",
          "funding_blocked"
        ],
        "title": "FundingState",
        "type": "object"
      },
      "Generate": {
        "additionalProperties": false,
        "properties": {
          "currencies": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Currencies",
            "type": "array"
          },
          "cycle_density": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cycle Density",
            "type": "integer"
          },
          "invoices": {
            "default": 30,
            "maximum": 20000.0,
            "minimum": 1.0,
            "title": "Invoices",
            "type": "integer"
          },
          "members": {
            "default": 6,
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Members",
            "type": "integer"
          },
          "seed": {
            "default": 1,
            "title": "Seed",
            "type": "integer"
          }
        },
        "title": "Generate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HoldView": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          },
          "member_name": {
            "title": "Member Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "member_id",
          "member_name",
          "currency",
          "amount_minor",
          "status"
        ],
        "title": "HoldView",
        "type": "object"
      },
...
      "JobView": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "member_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step": {
            "title": "Step",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "step",
          "member_id",
          "status",
          "attempts",
          "last_error",
          "result",
          "updated_at"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "LedgerLine": {
        "description": "One of the member's own ledger movements. The counterparty is always Mesh settlement.",
        "properties": {
          "account": {
            "enum": [
              "BALANCE",
              "HELD",
              "CARRIED"
            ],
            "title": "Account",
            "type": "string"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "counterparty": {
            "default": "Mesh settlement",
            "title": "Counterparty",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "entry_kind": {
            "title": "Entry Kind",
            "type": "string"
          }
        },
        "required": [
          "entry_kind",
          "account",
          "description",
          "amount",
          "created_at"
        ],
        "title": "LedgerLine",
        "type": "object"
      },
...
      "MemberSettlement": {
        "description": "The member's own side of settlement (MC-PRV-02): no other member is ever named.",
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "carried": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "counterparty": {
            "default": "Mesh settlement",
            "title": "Counterparty",
            "type": "string"
          },
          "held": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "instruction": {
            "enum": [
              "DEBIT",
              "CREDIT",
              "NONE"
            ],
            "title": "Instruction",
            "type": "string"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/LedgerLine"
            },
            "title": "Ledger",
            "type": "array"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "settled_by_netting": {
            "title": "Settled By Netting",
            "type": "integer"
          },
          "settled_by_transfer": {
            "title": "Settled By Transfer",
            "type": "integer"
          },
          "status": {
            "enum": [
              "PENDING",
              "FUNDS_HELD",
              "SETTLED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "instruction",
          "amount",
          "reference",
          "held",
          "carried",
          "settled_by_netting",
          "settled_by_transfer",
          "ledger"
        ],
        "title": "MemberSettlement",
        "type": "object"
      },
...
      "Money": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "ResetResult": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ResetResult",
        "type": "object"
      },
...
      "RunView": {
        "properties": {
          "approvals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Approvals",
            "type": "array"
          },
          "computations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Computations",
            "type": "array"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "current_attempt": {
            "title": "Current Attempt",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "own_positions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Own Positions",
            "type": "array"
          },
          "settlement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemberSettlement"
              },
              {
                "type": "null"
              }
            ]
          },
          "settlement_detail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SettlementDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "statement_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "id",
          "status",
          "current_attempt",
          "started_at"
        ],
        "title": "RunView",
        "type": "object"
      },
...
      "SettlementDetail": {
        "description": "Wise staff view of a run's settlement: holds, jobs, transfers and clearing.",
        "properties": {
          "clearing": {
            "items": {
              "$ref": "#/components/schemas/AccountBalance"
            },
            "title": "Clearing",
            "type": "array"
          },
          "commit_entry_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Entry Seq"
          },
          "holds": {
            "items": {
              "$ref": "#/components/schemas/HoldView"
            },
            "title": "Holds",
            "type": "array"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "title": "Jobs",
            "type": "array"
          },
          "transfers": {
            "items": {
              "$ref": "#/components/schemas/TransferView"
            },
            "title": "Transfers",
            "type": "array"
          }
        },
        "required": [
          "holds",
          "jobs",
          "transfers",
          "clearing",
          "commit_entry_seq"
        ],
        "title": "SettlementDetail",
        "type": "object"
      },
      "Simulate": {
        "additionalProperties": false,
        "properties": {
          "confirm_pct": {
            "default": 100,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confirm Pct",
            "type": "integer"
          },
          "dispute_pct": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Dispute Pct",
            "type": "integer"
          },
          "seed": {
            "default": 1,
            "title": "Seed",
            "type": "integer"
          }
        },
        "title": "Simulate",
        "type": "object"
      },
...
      "TransferView": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payer": {
            "title": "Payer",
            "type": "string"
          },
          "receiver": {
            "title": "Receiver",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "payer",
          "receiver",
          "currency",
          "amount_minor",
          "kind",
          "status"
        ],
        "title": "TransferView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multilateral invoice netting and settlement (hackathon prototype).",
    "title": "Wise Mesh Continental API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/demo/expire-approvals": {
      "post": {
        "description": "Remove members who have not approved, then recompute (the approval deadline).",
        "operationId": "expire_approvals_v1_demo_expire_approvals_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpireApprovals"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Expire Approvals",
        "tags": [
          "demo"
        ]
      }
    },
    "/v1/demo/funding-failure": {
      "post": {
        "description": "Mark a payer as unable to fund (or able again); the next prepare excludes it.",
        "operationId": "funding_failure_v1_demo_funding_failure_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundingFailure"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundingState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Funding Failure",
        "tags": [
          "demo"
        ]
      }
    },
    "/v1/demo/generate": {
      "post": {
        "operationId": "generate_v1_demo_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Generate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Generate V1 Demo Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "demo"
        ]
      }
    },
...
    "/v1/demo/reset": {
      "post": {
        "operationId": "reset_v1_demo_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset",
        "tags": [
          "demo"
        ]
      }
    },
    "/v1/demo/scenarios/{name}": {
      "post": {
        "operationId": "scenario_v1_demo_scenarios__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Scenario V1 Demo Scenarios  Name  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario",
        "tags": [
          "demo"
        ]
      }
    },
    "/v1/demo/simulate-confirmations": {
      "post": {
        "operationId": "simulate_v1_demo_simulate_confirmations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Simulate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Simulate V1 Demo Simulate Confirmations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}